Struct aws_sdk_ses::client::Client

pub struct Client<C = DynConnector, M = AwsMiddleware, R = Standard> { /* fields omitted */ }

An ergonomic service client for SimpleEmailService.

This client allows ergonomic access to a SimpleEmailService-shaped service. Each method corresponds to an endpoint defined in the service’s Smithy model, and the request and response shapes are auto-generated from that same model.

Using a Client

Once you have a client set up, you can access the service’s endpoints by calling the appropriate method on Client. Each such method returns a request builder for that endpoint, with methods for setting the various fields of the request. Once your request is complete, use the send method to send the request. send returns a future, which you then have to .await to get the service’s response.

Implementations

impl<C, M, R> Client<C, M, R>

pub fn with_config(client: Client<C, M, R>, conf: Config) -> Self

pub fn conf(&self) -> &Config

impl<C, M, R> Client<C, M, R> where
C: SmithyConnector,
M: SmithyMiddleware<C>,
R: NewRequestPolicy,

pub fn clone_receipt_rule_set(&self) -> CloneReceiptRuleSet<C, M, R>

pub fn create_configuration_set(&self) -> CreateConfigurationSet<C, M, R>

pub fn create_configuration_set_event_destination(
&self
) -> CreateConfigurationSetEventDestination<C, M, R>

pub fn create_configuration_set_tracking_options(
&self
) -> CreateConfigurationSetTrackingOptions<C, M, R>

pub fn create_custom_verification_email_template(
&self
) -> CreateCustomVerificationEmailTemplate<C, M, R>

pub fn create_receipt_filter(&self) -> CreateReceiptFilter<C, M, R>

pub fn create_receipt_rule(&self) -> CreateReceiptRule<C, M, R>

pub fn create_receipt_rule_set(&self) -> CreateReceiptRuleSet<C, M, R>

pub fn create_template(&self) -> CreateTemplate<C, M, R>

pub fn delete_configuration_set(&self) -> DeleteConfigurationSet<C, M, R>

pub fn delete_configuration_set_event_destination(
&self
) -> DeleteConfigurationSetEventDestination<C, M, R>

pub fn delete_configuration_set_tracking_options(
&self
) -> DeleteConfigurationSetTrackingOptions<C, M, R>

pub fn delete_custom_verification_email_template(
&self
) -> DeleteCustomVerificationEmailTemplate<C, M, R>

pub fn delete_identity(&self) -> DeleteIdentity<C, M, R>

pub fn delete_identity_policy(&self) -> DeleteIdentityPolicy<C, M, R>

pub fn delete_receipt_filter(&self) -> DeleteReceiptFilter<C, M, R>

pub fn delete_receipt_rule(&self) -> DeleteReceiptRule<C, M, R>

pub fn delete_receipt_rule_set(&self) -> DeleteReceiptRuleSet<C, M, R>

pub fn delete_template(&self) -> DeleteTemplate<C, M, R>

pub fn delete_verified_email_address(
&self
) -> DeleteVerifiedEmailAddress<C, M, R>

pub fn describe_active_receipt_rule_set(
&self
) -> DescribeActiveReceiptRuleSet<C, M, R>

pub fn describe_configuration_set(&self) -> DescribeConfigurationSet<C, M, R>

pub fn describe_receipt_rule(&self) -> DescribeReceiptRule<C, M, R>

pub fn describe_receipt_rule_set(&self) -> DescribeReceiptRuleSet<C, M, R>

pub fn get_account_sending_enabled(&self) -> GetAccountSendingEnabled<C, M, R>

pub fn get_custom_verification_email_template(
&self
) -> GetCustomVerificationEmailTemplate<C, M, R>

pub fn get_identity_dkim_attributes(&self) -> GetIdentityDkimAttributes<C, M, R>

pub fn get_identity_mail_from_domain_attributes(
&self
) -> GetIdentityMailFromDomainAttributes<C, M, R>

pub fn get_identity_notification_attributes(
&self
) -> GetIdentityNotificationAttributes<C, M, R>

pub fn get_identity_policies(&self) -> GetIdentityPolicies<C, M, R>

pub fn get_identity_verification_attributes(
&self
) -> GetIdentityVerificationAttributes<C, M, R>

pub fn get_send_quota(&self) -> GetSendQuota<C, M, R>

pub fn get_send_statistics(&self) -> GetSendStatistics<C, M, R>

pub fn get_template(&self) -> GetTemplate<C, M, R>

pub fn list_configuration_sets(&self) -> ListConfigurationSets<C, M, R>

pub fn list_custom_verification_email_templates(
&self
) -> ListCustomVerificationEmailTemplates<C, M, R>

pub fn list_identities(&self) -> ListIdentities<C, M, R>

pub fn list_identity_policies(&self) -> ListIdentityPolicies<C, M, R>

pub fn list_receipt_filters(&self) -> ListReceiptFilters<C, M, R>

pub fn list_receipt_rule_sets(&self) -> ListReceiptRuleSets<C, M, R>

pub fn list_templates(&self) -> ListTemplates<C, M, R>

pub fn list_verified_email_addresses(
&self
) -> ListVerifiedEmailAddresses<C, M, R>

pub fn put_configuration_set_delivery_options(
&self
) -> PutConfigurationSetDeliveryOptions<C, M, R>

pub fn put_identity_policy(&self) -> PutIdentityPolicy<C, M, R>

pub fn reorder_receipt_rule_set(&self) -> ReorderReceiptRuleSet<C, M, R>

pub fn send_bounce(&self) -> SendBounce<C, M, R>

pub fn send_bulk_templated_email(&self) -> SendBulkTemplatedEmail<C, M, R>

pub fn send_custom_verification_email(
&self
) -> SendCustomVerificationEmail<C, M, R>

pub fn send_email(&self) -> SendEmail<C, M, R>

pub fn send_raw_email(&self) -> SendRawEmail<C, M, R>

pub fn send_templated_email(&self) -> SendTemplatedEmail<C, M, R>

pub fn set_active_receipt_rule_set(&self) -> SetActiveReceiptRuleSet<C, M, R>

pub fn set_identity_dkim_enabled(&self) -> SetIdentityDkimEnabled<C, M, R>

pub fn set_identity_feedback_forwarding_enabled(
&self
) -> SetIdentityFeedbackForwardingEnabled<C, M, R>

pub fn set_identity_headers_in_notifications_enabled(
&self
) -> SetIdentityHeadersInNotificationsEnabled<C, M, R>

pub fn set_identity_mail_from_domain(
&self
) -> SetIdentityMailFromDomain<C, M, R>

pub fn set_identity_notification_topic(
&self
) -> SetIdentityNotificationTopic<C, M, R>

pub fn set_receipt_rule_position(&self) -> SetReceiptRulePosition<C, M, R>

pub fn test_render_template(&self) -> TestRenderTemplate<C, M, R>

pub fn update_account_sending_enabled(
&self
) -> UpdateAccountSendingEnabled<C, M, R>

pub fn update_configuration_set_event_destination(
&self
) -> UpdateConfigurationSetEventDestination<C, M, R>

pub fn update_configuration_set_reputation_metrics_enabled(
&self
) -> UpdateConfigurationSetReputationMetricsEnabled<C, M, R>

pub fn update_configuration_set_sending_enabled(
&self
) -> UpdateConfigurationSetSendingEnabled<C, M, R>

pub fn update_configuration_set_tracking_options(
&self
) -> UpdateConfigurationSetTrackingOptions<C, M, R>

pub fn update_custom_verification_email_template(
&self
) -> UpdateCustomVerificationEmailTemplate<C, M, R>

pub fn update_receipt_rule(&self) -> UpdateReceiptRule<C, M, R>

pub fn update_template(&self) -> UpdateTemplate<C, M, R>

pub fn verify_domain_dkim(&self) -> VerifyDomainDkim<C, M, R>

pub fn verify_domain_identity(&self) -> VerifyDomainIdentity<C, M, R>

pub fn verify_email_address(&self) -> VerifyEmailAddress<C, M, R>

pub fn verify_email_identity(&self) -> VerifyEmailIdentity<C, M, R>

impl<C> Client<C, AwsMiddleware, Standard>

pub fn from_conf_conn(conf: Config, conn: C) -> Self

impl Client<DynConnector, AwsMiddleware, Standard>

pub fn new(config: &Config) -> Self

pub fn from_conf(conf: Config) -> Self