Enum aws_sdk_ses::Error

#[non_exhaustive]
pub enum Error {

    AccountSendingPausedException(AccountSendingPausedException),
    AlreadyExistsException(AlreadyExistsException),
    CannotDeleteException(CannotDeleteException),
    ConfigurationSetAlreadyExistsException(ConfigurationSetAlreadyExistsException),
    ConfigurationSetDoesNotExistException(ConfigurationSetDoesNotExistException),
    ConfigurationSetSendingPausedException(ConfigurationSetSendingPausedException),
    CustomVerificationEmailInvalidContentException(CustomVerificationEmailInvalidContentException),
    CustomVerificationEmailTemplateAlreadyExistsException(CustomVerificationEmailTemplateAlreadyExistsException),
    CustomVerificationEmailTemplateDoesNotExistException(CustomVerificationEmailTemplateDoesNotExistException),
    EventDestinationAlreadyExistsException(EventDestinationAlreadyExistsException),
    EventDestinationDoesNotExistException(EventDestinationDoesNotExistException),
    FromEmailAddressNotVerifiedException(FromEmailAddressNotVerifiedException),
    InvalidCloudWatchDestinationException(InvalidCloudWatchDestinationException),
    InvalidConfigurationSetException(InvalidConfigurationSetException),
    InvalidDeliveryOptionsException(InvalidDeliveryOptionsException),
    InvalidFirehoseDestinationException(InvalidFirehoseDestinationException),
    InvalidLambdaFunctionException(InvalidLambdaFunctionException),
    InvalidPolicyException(InvalidPolicyException),
    InvalidRenderingParameterException(InvalidRenderingParameterException),
    InvalidS3ConfigurationException(InvalidS3ConfigurationException),
    InvalidSnsDestinationException(InvalidSnsDestinationException),
    InvalidSnsTopicException(InvalidSnsTopicException),
    InvalidTemplateException(InvalidTemplateException),
    InvalidTrackingOptionsException(InvalidTrackingOptionsException),
    LimitExceededException(LimitExceededException),
    MailFromDomainNotVerifiedException(MailFromDomainNotVerifiedException),
    MessageRejected(MessageRejected),
    MissingRenderingAttributeException(MissingRenderingAttributeException),
    ProductionAccessNotGrantedException(ProductionAccessNotGrantedException),
    RuleDoesNotExistException(RuleDoesNotExistException),
    RuleSetDoesNotExistException(RuleSetDoesNotExistException),
    TemplateDoesNotExistException(TemplateDoesNotExistException),
    TrackingOptionsAlreadyExistsException(TrackingOptionsAlreadyExistsException),
    TrackingOptionsDoesNotExistException(TrackingOptionsDoesNotExistException),
    Unhandled(Box<dyn Error + Send + Sync + 'static>),

}

All possible error types for this service.

Variants (Non-exhaustive)

Non-exhaustive enums could have additional variants added in future. Therefore, when matching against variants of non-exhaustive enums, an extra wildcard arm must be added to account for any future variants.

AccountSendingPausedException(AccountSendingPausedException)

Indicates that email sending is disabled for your entire Amazon SES account.

You can enable or disable email sending for your Amazon SES account using UpdateAccountSendingEnabled.

Tuple Fields of AccountSendingPausedException

0: AccountSendingPausedException

AlreadyExistsException(AlreadyExistsException)

Indicates that a resource could not be created because of a naming conflict.

Tuple Fields of AlreadyExistsException

0: AlreadyExistsException

CannotDeleteException(CannotDeleteException)

Indicates that the delete operation could not be completed.

Tuple Fields of CannotDeleteException

0: CannotDeleteException

ConfigurationSetAlreadyExistsException(ConfigurationSetAlreadyExistsException)

Indicates that the configuration set could not be created because of a naming conflict.

Tuple Fields of ConfigurationSetAlreadyExistsException

0: ConfigurationSetAlreadyExistsException

ConfigurationSetDoesNotExistException(ConfigurationSetDoesNotExistException)

Indicates that the configuration set does not exist.

Tuple Fields of ConfigurationSetDoesNotExistException

0: ConfigurationSetDoesNotExistException

ConfigurationSetSendingPausedException(ConfigurationSetSendingPausedException)

Indicates that email sending is disabled for the configuration set.

You can enable or disable email sending for a configuration set using UpdateConfigurationSetSendingEnabled.

Tuple Fields of ConfigurationSetSendingPausedException

0: ConfigurationSetSendingPausedException

CustomVerificationEmailInvalidContentException(CustomVerificationEmailInvalidContentException)

Indicates that custom verification email template provided content is invalid.

Tuple Fields of CustomVerificationEmailInvalidContentException

0: CustomVerificationEmailInvalidContentException

CustomVerificationEmailTemplateAlreadyExistsException(CustomVerificationEmailTemplateAlreadyExistsException)

Indicates that a custom verification email template with the name you specified already exists.

Tuple Fields of CustomVerificationEmailTemplateAlreadyExistsException

0: CustomVerificationEmailTemplateAlreadyExistsException

CustomVerificationEmailTemplateDoesNotExistException(CustomVerificationEmailTemplateDoesNotExistException)

Indicates that a custom verification email template with the name you specified does not exist.

Tuple Fields of CustomVerificationEmailTemplateDoesNotExistException

0: CustomVerificationEmailTemplateDoesNotExistException

EventDestinationAlreadyExistsException(EventDestinationAlreadyExistsException)

Indicates that the event destination could not be created because of a naming conflict.

Tuple Fields of EventDestinationAlreadyExistsException

0: EventDestinationAlreadyExistsException

EventDestinationDoesNotExistException(EventDestinationDoesNotExistException)

Indicates that the event destination does not exist.

Tuple Fields of EventDestinationDoesNotExistException

0: EventDestinationDoesNotExistException

FromEmailAddressNotVerifiedException(FromEmailAddressNotVerifiedException)

Indicates that the sender address specified for a custom verification email is not verified, and is therefore not eligible to send the custom verification email.

Tuple Fields of FromEmailAddressNotVerifiedException

0: FromEmailAddressNotVerifiedException

InvalidCloudWatchDestinationException(InvalidCloudWatchDestinationException)

Indicates that the Amazon CloudWatch destination is invalid. See the error message for details.

Tuple Fields of InvalidCloudWatchDestinationException

0: InvalidCloudWatchDestinationException

InvalidConfigurationSetException(InvalidConfigurationSetException)

Indicates that the configuration set is invalid. See the error message for details.

Tuple Fields of InvalidConfigurationSetException

0: InvalidConfigurationSetException

InvalidDeliveryOptionsException(InvalidDeliveryOptionsException)

Indicates that provided delivery option is invalid.

Tuple Fields of InvalidDeliveryOptionsException

0: InvalidDeliveryOptionsException

InvalidFirehoseDestinationException(InvalidFirehoseDestinationException)

Indicates that the Amazon Kinesis Firehose destination is invalid. See the error message for details.

Tuple Fields of InvalidFirehoseDestinationException

0: InvalidFirehoseDestinationException

InvalidLambdaFunctionException(InvalidLambdaFunctionException)

Indicates that the provided AWS Lambda function is invalid, or that Amazon SES could not execute the provided function, possibly due to permissions issues. For information about giving permissions, see the Amazon SES Developer Guide.

Tuple Fields of InvalidLambdaFunctionException

0: InvalidLambdaFunctionException

InvalidPolicyException(InvalidPolicyException)

Indicates that the provided policy is invalid. Check the error stack for more information about what caused the error.

Tuple Fields of InvalidPolicyException

0: InvalidPolicyException

InvalidRenderingParameterException(InvalidRenderingParameterException)

Indicates that one or more of the replacement values you provided is invalid. This error may occur when the TemplateData object contains invalid JSON.

Tuple Fields of InvalidRenderingParameterException

0: InvalidRenderingParameterException

InvalidS3ConfigurationException(InvalidS3ConfigurationException)

Indicates that the provided Amazon S3 bucket or AWS KMS encryption key is invalid, or that Amazon SES could not publish to the bucket, possibly due to permissions issues. For information about giving permissions, see the Amazon SES Developer Guide.

Tuple Fields of InvalidS3ConfigurationException

0: InvalidS3ConfigurationException

InvalidSnsDestinationException(InvalidSnsDestinationException)

Indicates that the Amazon Simple Notification Service (Amazon SNS) destination is invalid. See the error message for details.

Tuple Fields of InvalidSnsDestinationException

0: InvalidSnsDestinationException

InvalidSnsTopicException(InvalidSnsTopicException)

Indicates that the provided Amazon SNS topic is invalid, or that Amazon SES could not publish to the topic, possibly due to permissions issues. For information about giving permissions, see the Amazon SES Developer Guide.

Tuple Fields of InvalidSnsTopicException

0: InvalidSnsTopicException

InvalidTemplateException(InvalidTemplateException)

Indicates that the template that you specified could not be rendered. This issue may occur when a template refers to a partial that does not exist.

Tuple Fields of InvalidTemplateException

0: InvalidTemplateException

InvalidTrackingOptionsException(InvalidTrackingOptionsException)

Indicates that the custom domain to be used for open and click tracking redirects is invalid. This error appears most often in the following situations:

When the tracking domain you specified is not verified in Amazon SES.

When the tracking domain you specified is not a valid domain or subdomain.

Tuple Fields of InvalidTrackingOptionsException

0: InvalidTrackingOptionsException

LimitExceededException(LimitExceededException)

Indicates that a resource could not be created because of service limits. For a list of Amazon SES limits, see the Amazon SES Developer Guide.

Tuple Fields of LimitExceededException

0: LimitExceededException

MailFromDomainNotVerifiedException(MailFromDomainNotVerifiedException)

Indicates that the message could not be sent because Amazon SES could not read the MX record required to use the specified MAIL FROM domain. For information about editing the custom MAIL FROM domain settings for an identity, see the Amazon SES Developer Guide.

Tuple Fields of MailFromDomainNotVerifiedException

0: MailFromDomainNotVerifiedException

MessageRejected(MessageRejected)

Indicates that the action failed, and the message could not be sent. Check the error stack for more information about what caused the error.

Tuple Fields of MessageRejected

0: MessageRejected

MissingRenderingAttributeException(MissingRenderingAttributeException)

Indicates that one or more of the replacement values for the specified template was not specified. Ensure that the TemplateData object contains references to all of the replacement tags in the specified template.

Tuple Fields of MissingRenderingAttributeException

0: MissingRenderingAttributeException

ProductionAccessNotGrantedException(ProductionAccessNotGrantedException)

Indicates that the account has not been granted production access.

Tuple Fields of ProductionAccessNotGrantedException

0: ProductionAccessNotGrantedException

RuleDoesNotExistException(RuleDoesNotExistException)

Indicates that the provided receipt rule does not exist.

Tuple Fields of RuleDoesNotExistException

0: RuleDoesNotExistException

RuleSetDoesNotExistException(RuleSetDoesNotExistException)

Indicates that the provided receipt rule set does not exist.

Tuple Fields of RuleSetDoesNotExistException

0: RuleSetDoesNotExistException

TemplateDoesNotExistException(TemplateDoesNotExistException)

Indicates that the Template object you specified does not exist in your Amazon SES account.

Tuple Fields of TemplateDoesNotExistException

0: TemplateDoesNotExistException

TrackingOptionsAlreadyExistsException(TrackingOptionsAlreadyExistsException)

Indicates that the configuration set you specified already contains a TrackingOptions object.

Tuple Fields of TrackingOptionsAlreadyExistsException

0: TrackingOptionsAlreadyExistsException

TrackingOptionsDoesNotExistException(TrackingOptionsDoesNotExistException)

Indicates that the TrackingOptions object you specified does not exist.

Tuple Fields of TrackingOptionsDoesNotExistException

0: TrackingOptionsDoesNotExistException

Unhandled(Box<dyn Error + Send + Sync + 'static>)

An unhandled error occurred.

Tuple Fields of Unhandled

0: Box<dyn Error + Send + Sync + 'static>