Module aws_sdk_ses::error

Modules

account_sending_paused_exception

See AccountSendingPausedException

already_exists_exception

See AlreadyExistsException

cannot_delete_exception

See CannotDeleteException

configuration_set_already_exists_exception

See ConfigurationSetAlreadyExistsException

configuration_set_does_not_exist_exception

See ConfigurationSetDoesNotExistException

configuration_set_sending_paused_exception

See ConfigurationSetSendingPausedException

custom_verification_email_invalid_content_exception

See CustomVerificationEmailInvalidContentException

custom_verification_email_template_already_exists_exception

See CustomVerificationEmailTemplateAlreadyExistsException

custom_verification_email_template_does_not_exist_exception

See CustomVerificationEmailTemplateDoesNotExistException

event_destination_already_exists_exception

See EventDestinationAlreadyExistsException

event_destination_does_not_exist_exception

See EventDestinationDoesNotExistException

from_email_address_not_verified_exception

See FromEmailAddressNotVerifiedException

invalid_cloud_watch_destination_exception

See InvalidCloudWatchDestinationException

invalid_configuration_set_exception

See InvalidConfigurationSetException

invalid_delivery_options_exception

See InvalidDeliveryOptionsException

invalid_firehose_destination_exception

See InvalidFirehoseDestinationException

invalid_lambda_function_exception

See InvalidLambdaFunctionException

invalid_policy_exception

See InvalidPolicyException

invalid_rendering_parameter_exception

See InvalidRenderingParameterException

invalid_s3_configuration_exception

See InvalidS3ConfigurationException

invalid_sns_destination_exception

See InvalidSnsDestinationException

invalid_sns_topic_exception

See InvalidSnsTopicException

invalid_template_exception

See InvalidTemplateException

invalid_tracking_options_exception

See InvalidTrackingOptionsException

limit_exceeded_exception

See LimitExceededException

mail_from_domain_not_verified_exception

See MailFromDomainNotVerifiedException

message_rejected

See MessageRejected

missing_rendering_attribute_exception

See MissingRenderingAttributeException

production_access_not_granted_exception

See ProductionAccessNotGrantedException

rule_does_not_exist_exception

See RuleDoesNotExistException

rule_set_does_not_exist_exception

See RuleSetDoesNotExistException

template_does_not_exist_exception

See TemplateDoesNotExistException

tracking_options_already_exists_exception

See TrackingOptionsAlreadyExistsException

tracking_options_does_not_exist_exception

See TrackingOptionsDoesNotExistException

Structs

AccountSendingPausedException

Indicates that email sending is disabled for your entire Amazon SES account.

AlreadyExistsException

Indicates that a resource could not be created because of a naming conflict.

CannotDeleteException

Indicates that the delete operation could not be completed.

CloneReceiptRuleSetError

ConfigurationSetAlreadyExistsException

Indicates that the configuration set could not be created because of a naming

ConfigurationSetDoesNotExistException

Indicates that the configuration set does not exist.

ConfigurationSetSendingPausedException

Indicates that email sending is disabled for the configuration set.

CreateConfigurationSetError

CreateConfigurationSetEventDestinationError

CreateConfigurationSetTrackingOptionsError

CreateCustomVerificationEmailTemplateError

CreateReceiptFilterError

CreateReceiptRuleError

CreateReceiptRuleSetError

CreateTemplateError

CustomVerificationEmailInvalidContentException

Indicates that custom verification email template provided content is invalid.

CustomVerificationEmailTemplateAlreadyExistsException

Indicates that a custom verification email template with the name you specified

CustomVerificationEmailTemplateDoesNotExistException

Indicates that a custom verification email template with the name you specified does

DeleteConfigurationSetError

DeleteConfigurationSetEventDestinationError

DeleteConfigurationSetTrackingOptionsError

DeleteCustomVerificationEmailTemplateError

DeleteIdentityError

DeleteIdentityPolicyError

DeleteReceiptFilterError

DeleteReceiptRuleError

DeleteReceiptRuleSetError

DeleteTemplateError

DeleteVerifiedEmailAddressError

DescribeActiveReceiptRuleSetError

DescribeConfigurationSetError

DescribeReceiptRuleError

DescribeReceiptRuleSetError

EventDestinationAlreadyExistsException

Indicates that the event destination could not be created because of a naming

EventDestinationDoesNotExistException

Indicates that the event destination does not exist.

FromEmailAddressNotVerifiedException

Indicates that the sender address specified for a custom verification email is not

GetAccountSendingEnabledError

GetCustomVerificationEmailTemplateError

GetIdentityDkimAttributesError

GetIdentityMailFromDomainAttributesError

GetIdentityNotificationAttributesError

GetIdentityPoliciesError

GetIdentityVerificationAttributesError

GetSendQuotaError

GetSendStatisticsError

GetTemplateError

InvalidCloudWatchDestinationException

Indicates that the Amazon CloudWatch destination is invalid. See the error message for

InvalidConfigurationSetException

Indicates that the configuration set is invalid. See the error message for

InvalidDeliveryOptionsException

Indicates that provided delivery option is invalid.

InvalidFirehoseDestinationException

Indicates that the Amazon Kinesis Firehose destination is invalid. See the error

InvalidLambdaFunctionException

Indicates that the provided AWS Lambda function is invalid, or that Amazon SES could

InvalidPolicyException

Indicates that the provided policy is invalid. Check the error stack for more

InvalidRenderingParameterException

Indicates that one or more of the replacement values you provided is invalid. This

InvalidS3ConfigurationException

Indicates that the provided Amazon S3 bucket or AWS KMS encryption key is invalid, or

InvalidSnsDestinationException

Indicates that the Amazon Simple Notification Service (Amazon SNS) destination is

InvalidSnsTopicException

Indicates that the provided Amazon SNS topic is invalid, or that Amazon SES could not

InvalidTemplateException

Indicates that the template that you specified could not be rendered. This issue may

InvalidTrackingOptionsException

Indicates that the custom domain to be used for open and click tracking redirects is

LimitExceededException

Indicates that a resource could not be created because of service limits. For a list

ListConfigurationSetsError

ListCustomVerificationEmailTemplatesError

ListIdentitiesError

ListIdentityPoliciesError

ListReceiptFiltersError

ListReceiptRuleSetsError

ListTemplatesError

ListVerifiedEmailAddressesError

MailFromDomainNotVerifiedException

Indicates that the message could not be sent because Amazon SES could not read the MX

MessageRejected

Indicates that the action failed, and the message could not be sent. Check the error

MissingRenderingAttributeException

Indicates that one or more of the replacement values for the specified template was

ProductionAccessNotGrantedException

Indicates that the account has not been granted production access.

PutConfigurationSetDeliveryOptionsError

PutIdentityPolicyError

ReorderReceiptRuleSetError

RuleDoesNotExistException

Indicates that the provided receipt rule does not exist.

RuleSetDoesNotExistException

Indicates that the provided receipt rule set does not exist.

SendBounceError

SendBulkTemplatedEmailError

SendCustomVerificationEmailError

SendEmailError

SendRawEmailError

SendTemplatedEmailError

SetActiveReceiptRuleSetError

SetIdentityDkimEnabledError

SetIdentityFeedbackForwardingEnabledError

SetIdentityHeadersInNotificationsEnabledError

SetIdentityMailFromDomainError

SetIdentityNotificationTopicError

SetReceiptRulePositionError

TemplateDoesNotExistException

Indicates that the Template object you specified does not exist in your Amazon SES

TestRenderTemplateError

TrackingOptionsAlreadyExistsException

Indicates that the configuration set you specified already contains a TrackingOptions

TrackingOptionsDoesNotExistException

Indicates that the TrackingOptions object you specified does not exist.

UpdateAccountSendingEnabledError

UpdateConfigurationSetEventDestinationError

UpdateConfigurationSetReputationMetricsEnabledError

UpdateConfigurationSetSendingEnabledError

UpdateConfigurationSetTrackingOptionsError

UpdateCustomVerificationEmailTemplateError

UpdateReceiptRuleError

UpdateTemplateError

VerifyDomainDkimError

VerifyDomainIdentityError

VerifyEmailAddressError

VerifyEmailIdentityError

Enums

CloneReceiptRuleSetErrorKind

CreateConfigurationSetErrorKind

CreateConfigurationSetEventDestinationErrorKind

CreateConfigurationSetTrackingOptionsErrorKind

CreateCustomVerificationEmailTemplateErrorKind

CreateReceiptFilterErrorKind

CreateReceiptRuleErrorKind

CreateReceiptRuleSetErrorKind

CreateTemplateErrorKind

DeleteConfigurationSetErrorKind

DeleteConfigurationSetEventDestinationErrorKind

DeleteConfigurationSetTrackingOptionsErrorKind

DeleteCustomVerificationEmailTemplateErrorKind

DeleteIdentityErrorKind

DeleteIdentityPolicyErrorKind

DeleteReceiptFilterErrorKind

DeleteReceiptRuleErrorKind

DeleteReceiptRuleSetErrorKind

DeleteTemplateErrorKind

DeleteVerifiedEmailAddressErrorKind

DescribeActiveReceiptRuleSetErrorKind

DescribeConfigurationSetErrorKind

DescribeReceiptRuleErrorKind

DescribeReceiptRuleSetErrorKind

GetAccountSendingEnabledErrorKind

GetCustomVerificationEmailTemplateErrorKind

GetIdentityDkimAttributesErrorKind

GetIdentityMailFromDomainAttributesErrorKind

GetIdentityNotificationAttributesErrorKind

GetIdentityPoliciesErrorKind

GetIdentityVerificationAttributesErrorKind

GetSendQuotaErrorKind

GetSendStatisticsErrorKind

GetTemplateErrorKind

ListConfigurationSetsErrorKind

ListCustomVerificationEmailTemplatesErrorKind

ListIdentitiesErrorKind

ListIdentityPoliciesErrorKind

ListReceiptFiltersErrorKind

ListReceiptRuleSetsErrorKind

ListTemplatesErrorKind

ListVerifiedEmailAddressesErrorKind

PutConfigurationSetDeliveryOptionsErrorKind

PutIdentityPolicyErrorKind

ReorderReceiptRuleSetErrorKind

SendBounceErrorKind

SendBulkTemplatedEmailErrorKind

SendCustomVerificationEmailErrorKind

SendEmailErrorKind

SendRawEmailErrorKind

SendTemplatedEmailErrorKind

SetActiveReceiptRuleSetErrorKind

SetIdentityDkimEnabledErrorKind

SetIdentityFeedbackForwardingEnabledErrorKind

SetIdentityHeadersInNotificationsEnabledErrorKind

SetIdentityMailFromDomainErrorKind

SetIdentityNotificationTopicErrorKind

SetReceiptRulePositionErrorKind

TestRenderTemplateErrorKind

UpdateAccountSendingEnabledErrorKind

UpdateConfigurationSetEventDestinationErrorKind

UpdateConfigurationSetReputationMetricsEnabledErrorKind

UpdateConfigurationSetSendingEnabledErrorKind

UpdateConfigurationSetTrackingOptionsErrorKind

UpdateCustomVerificationEmailTemplateErrorKind

UpdateReceiptRuleErrorKind

UpdateTemplateErrorKind

VerifyDomainDkimErrorKind

VerifyDomainIdentityErrorKind

VerifyEmailAddressErrorKind

VerifyEmailIdentityErrorKind