Module aws_sdk_ses::input

Modules

clone_receipt_rule_set_input

See CloneReceiptRuleSetInput

create_configuration_set_event_destination_input

See CreateConfigurationSetEventDestinationInput

create_configuration_set_input

See CreateConfigurationSetInput

create_configuration_set_tracking_options_input

See CreateConfigurationSetTrackingOptionsInput

create_custom_verification_email_template_input

See CreateCustomVerificationEmailTemplateInput

create_receipt_filter_input

See CreateReceiptFilterInput

create_receipt_rule_input

See CreateReceiptRuleInput

create_receipt_rule_set_input

See CreateReceiptRuleSetInput

create_template_input

See CreateTemplateInput

delete_configuration_set_event_destination_input

See DeleteConfigurationSetEventDestinationInput

delete_configuration_set_input

See DeleteConfigurationSetInput

delete_configuration_set_tracking_options_input

See DeleteConfigurationSetTrackingOptionsInput

delete_custom_verification_email_template_input

See DeleteCustomVerificationEmailTemplateInput

delete_identity_input

See DeleteIdentityInput

delete_identity_policy_input

See DeleteIdentityPolicyInput

delete_receipt_filter_input

See DeleteReceiptFilterInput

delete_receipt_rule_input

See DeleteReceiptRuleInput

delete_receipt_rule_set_input

See DeleteReceiptRuleSetInput

delete_template_input

See DeleteTemplateInput

delete_verified_email_address_input

See DeleteVerifiedEmailAddressInput

describe_active_receipt_rule_set_input

See DescribeActiveReceiptRuleSetInput

describe_configuration_set_input

See DescribeConfigurationSetInput

describe_receipt_rule_input

See DescribeReceiptRuleInput

describe_receipt_rule_set_input

See DescribeReceiptRuleSetInput

get_account_sending_enabled_input

See GetAccountSendingEnabledInput

get_custom_verification_email_template_input

See GetCustomVerificationEmailTemplateInput

get_identity_dkim_attributes_input

See GetIdentityDkimAttributesInput

get_identity_mail_from_domain_attributes_input

See GetIdentityMailFromDomainAttributesInput

get_identity_notification_attributes_input

See GetIdentityNotificationAttributesInput

get_identity_policies_input

See GetIdentityPoliciesInput

get_identity_verification_attributes_input

See GetIdentityVerificationAttributesInput

get_send_quota_input

See GetSendQuotaInput

get_send_statistics_input

See GetSendStatisticsInput

get_template_input

See GetTemplateInput

list_configuration_sets_input

See ListConfigurationSetsInput

list_custom_verification_email_templates_input

See ListCustomVerificationEmailTemplatesInput

list_identities_input

See ListIdentitiesInput

list_identity_policies_input

See ListIdentityPoliciesInput

list_receipt_filters_input

See ListReceiptFiltersInput

list_receipt_rule_sets_input

See ListReceiptRuleSetsInput

list_templates_input

See ListTemplatesInput

list_verified_email_addresses_input

See ListVerifiedEmailAddressesInput

put_configuration_set_delivery_options_input

See PutConfigurationSetDeliveryOptionsInput

put_identity_policy_input

See PutIdentityPolicyInput

reorder_receipt_rule_set_input

See ReorderReceiptRuleSetInput

send_bounce_input

See SendBounceInput

send_bulk_templated_email_input

See SendBulkTemplatedEmailInput

send_custom_verification_email_input

See SendCustomVerificationEmailInput

send_email_input

See SendEmailInput

send_raw_email_input

See SendRawEmailInput

send_templated_email_input

See SendTemplatedEmailInput

set_active_receipt_rule_set_input

See SetActiveReceiptRuleSetInput

set_identity_dkim_enabled_input

See SetIdentityDkimEnabledInput

set_identity_feedback_forwarding_enabled_input

See SetIdentityFeedbackForwardingEnabledInput

set_identity_headers_in_notifications_enabled_input

See SetIdentityHeadersInNotificationsEnabledInput

set_identity_mail_from_domain_input

See SetIdentityMailFromDomainInput

set_identity_notification_topic_input

See SetIdentityNotificationTopicInput

set_receipt_rule_position_input

See SetReceiptRulePositionInput

test_render_template_input

See TestRenderTemplateInput

update_account_sending_enabled_input

See UpdateAccountSendingEnabledInput

update_configuration_set_event_destination_input

See UpdateConfigurationSetEventDestinationInput

update_configuration_set_reputation_metrics_enabled_input

See UpdateConfigurationSetReputationMetricsEnabledInput

update_configuration_set_sending_enabled_input

See UpdateConfigurationSetSendingEnabledInput

update_configuration_set_tracking_options_input

See UpdateConfigurationSetTrackingOptionsInput

update_custom_verification_email_template_input

See UpdateCustomVerificationEmailTemplateInput

update_receipt_rule_input

See UpdateReceiptRuleInput

update_template_input

See UpdateTemplateInput

verify_domain_dkim_input

See VerifyDomainDkimInput

verify_domain_identity_input

See VerifyDomainIdentityInput

verify_email_address_input

See VerifyEmailAddressInput

verify_email_identity_input

See VerifyEmailIdentityInput

Structs

CloneReceiptRuleSetInput

Represents a request to create a receipt rule set by cloning an existing one. You use

CreateConfigurationSetEventDestinationInput

Represents a request to create a configuration set event destination. A configuration

CreateConfigurationSetInput

Represents a request to create a configuration set. Configuration sets enable you to

CreateConfigurationSetTrackingOptionsInput

Represents a request to create an open and click tracking option object in a

CreateCustomVerificationEmailTemplateInput

Represents a request to create a custom verification email template.

CreateReceiptFilterInput

Represents a request to create a new IP address filter. You use IP address filters

CreateReceiptRuleInput

Represents a request to create a receipt rule. You use receipt rules to receive email

CreateReceiptRuleSetInput

Represents a request to create an empty receipt rule set. You use receipt rule sets to

CreateTemplateInput

Represents a request to create an email template. For more information, see the Amazon SES Developer

DeleteConfigurationSetEventDestinationInput

Represents a request to delete a configuration set event destination. Configuration

DeleteConfigurationSetInput

Represents a request to delete a configuration set. Configuration sets enable you to

DeleteConfigurationSetTrackingOptionsInput

Represents a request to delete open and click tracking options in a configuration set.

DeleteCustomVerificationEmailTemplateInput

Represents a request to delete an existing custom verification email template.

DeleteIdentityInput

Represents a request to delete one of your Amazon SES identities (an email address or

DeleteIdentityPolicyInput

Represents a request to delete a sending authorization policy for an identity. Sending

DeleteReceiptFilterInput

Represents a request to delete an IP address filter. You use IP address filters when

DeleteReceiptRuleInput

Represents a request to delete a receipt rule. You use receipt rules to receive email

DeleteReceiptRuleSetInput

Represents a request to delete a receipt rule set and all of the receipt rules it

DeleteTemplateInput

Represents a request to delete an email template. For more information, see the Amazon SES Developer

DeleteVerifiedEmailAddressInput

Represents a request to delete an email address from the list of email addresses you

DescribeActiveReceiptRuleSetInput

Represents a request to return the metadata and receipt rules for the receipt rule set

DescribeConfigurationSetInput

Represents a request to return the details of a configuration set. Configuration sets

DescribeReceiptRuleInput

Represents a request to return the details of a receipt rule. You use receipt rules to

DescribeReceiptRuleSetInput

Represents a request to return the details of a receipt rule set. You use receipt rule

GetAccountSendingEnabledInput

GetCustomVerificationEmailTemplateInput

Represents a request to retrieve an existing custom verification email

GetIdentityDkimAttributesInput

Represents a request for the status of Amazon SES Easy DKIM signing for an identity. For

GetIdentityMailFromDomainAttributesInput

Represents a request to return the Amazon SES custom MAIL FROM attributes for a list of

GetIdentityNotificationAttributesInput

Represents a request to return the notification attributes for a list of identities

GetIdentityPoliciesInput

Represents a request to return the requested sending authorization policies for an

GetIdentityVerificationAttributesInput

Represents a request to return the Amazon SES verification status of a list of identities.

GetSendQuotaInput

GetSendStatisticsInput

GetTemplateInput

ListConfigurationSetsInput

Represents a request to list the configuration sets associated with your AWS account.

ListCustomVerificationEmailTemplatesInput

Represents a request to list the existing custom verification email templates for your

ListIdentitiesInput

Represents a request to return a list of all identities (email addresses and domains)

ListIdentityPoliciesInput

Represents a request to return a list of sending authorization policies that are

ListReceiptFiltersInput

Represents a request to list the IP address filters that exist under your AWS account.

ListReceiptRuleSetsInput

Represents a request to list the receipt rule sets that exist under your AWS account.

ListTemplatesInput

ListVerifiedEmailAddressesInput

PutConfigurationSetDeliveryOptionsInput

A request to modify the delivery options for a configuration set.

PutIdentityPolicyInput

Represents a request to add or update a sending authorization policy for an identity.

ReorderReceiptRuleSetInput

Represents a request to reorder the receipt rules within a receipt rule set. You use

SendBounceInput

Represents a request to send a bounce message to the sender of an email you received

SendBulkTemplatedEmailInput

Represents a request to send a templated email to multiple destinations using Amazon SES.

SendCustomVerificationEmailInput

Represents a request to send a custom verification email to a specified

SendEmailInput

Represents a request to send a single formatted email using Amazon SES. For more

SendRawEmailInput

Represents a request to send a single raw email using Amazon SES. For more information, see

SendTemplatedEmailInput

Represents a request to send a templated email using Amazon SES. For more information, see

SetActiveReceiptRuleSetInput

Represents a request to set a receipt rule set as the active receipt rule set. You use

SetIdentityDkimEnabledInput

Represents a request to enable or disable Amazon SES Easy DKIM signing for an identity. For

SetIdentityFeedbackForwardingEnabledInput

Represents a request to enable or disable whether Amazon SES forwards you bounce and

SetIdentityHeadersInNotificationsEnabledInput

Represents a request to set whether Amazon SES includes the original email headers in the

SetIdentityMailFromDomainInput

Represents a request to enable or disable the Amazon SES custom MAIL FROM domain setup for

SetIdentityNotificationTopicInput

Represents a request to specify the Amazon SNS topic to which Amazon SES will publish bounce,

SetReceiptRulePositionInput

Represents a request to set the position of a receipt rule in a receipt rule set. You

TestRenderTemplateInput

UpdateAccountSendingEnabledInput

Represents a request to enable or disable the email sending capabilities for your

UpdateConfigurationSetEventDestinationInput

Represents a request to update the event destination of a configuration set.

UpdateConfigurationSetReputationMetricsEnabledInput

Represents a request to modify the reputation metric publishing settings for a

UpdateConfigurationSetSendingEnabledInput

Represents a request to enable or disable the email sending capabilities for a

UpdateConfigurationSetTrackingOptionsInput

Represents a request to update the tracking options for a configuration set.

UpdateCustomVerificationEmailTemplateInput

Represents a request to update an existing custom verification email template.

UpdateReceiptRuleInput

Represents a request to update a receipt rule. You use receipt rules to receive email

UpdateTemplateInput

VerifyDomainDkimInput

Represents a request to generate the CNAME records needed to set up Easy DKIM with

VerifyDomainIdentityInput

Represents a request to begin Amazon SES domain verification and to generate the TXT

VerifyEmailAddressInput

Represents a request to begin email address verification with Amazon SES. For information

VerifyEmailIdentityInput

Represents a request to begin email address verification with Amazon SES. For information