Module aws_sdk_ses::model

Modules

add_header_action

See AddHeaderAction

body

See Body

bounce_action

See BounceAction

bounced_recipient_info

See BouncedRecipientInfo

bulk_email_destination

See BulkEmailDestination

bulk_email_destination_status

See BulkEmailDestinationStatus

cloud_watch_destination

See CloudWatchDestination

cloud_watch_dimension_configuration

See CloudWatchDimensionConfiguration

configuration_set

See ConfigurationSet

content

See Content

custom_verification_email_template

See CustomVerificationEmailTemplate

delivery_options

See DeliveryOptions

destination

See Destination

event_destination

See EventDestination

extension_field

See ExtensionField

identity_dkim_attributes

See IdentityDkimAttributes

identity_mail_from_domain_attributes

See IdentityMailFromDomainAttributes

identity_notification_attributes

See IdentityNotificationAttributes

identity_verification_attributes

See IdentityVerificationAttributes

kinesis_firehose_destination

See KinesisFirehoseDestination

lambda_action

See LambdaAction

message

See Message

message_dsn

See MessageDsn

message_tag

See MessageTag

raw_message

See RawMessage

receipt_action

See ReceiptAction

receipt_filter

See ReceiptFilter

receipt_ip_filter

See ReceiptIpFilter

receipt_rule

See ReceiptRule

receipt_rule_set_metadata

See ReceiptRuleSetMetadata

recipient_dsn_fields

See RecipientDsnFields

reputation_options

See ReputationOptions

s3_action

See S3Action

send_data_point

See SendDataPoint

sns_action

See SnsAction

sns_destination

See SnsDestination

stop_action

See StopAction

template

See Template

template_metadata

See TemplateMetadata

tracking_options

See TrackingOptions

workmail_action

See WorkmailAction

Structs

AddHeaderAction

When included in a receipt rule, this action adds a header to the received

Body

Represents the body of the message. You can specify text, HTML, or both. If you use

BounceAction

When included in a receipt rule, this action rejects the received email by returning a

BouncedRecipientInfo

Recipient-related information to include in the Delivery Status Notification (DSN)

BulkEmailDestination

An array that contains one or more Destinations, as well as the tags and replacement

BulkEmailDestinationStatus

An object that contains the response from the SendBulkTemplatedEmail

CloudWatchDestination

Contains information associated with an Amazon CloudWatch event destination to which email

CloudWatchDimensionConfiguration

Contains the dimension configuration to use when you publish email sending events to

ConfigurationSet

The name of the configuration set.

Content

Represents textual data, plus an optional character set specification.

CustomVerificationEmailTemplate

Contains information about a custom verification email template.

DeliveryOptions

Specifies whether messages that use the configuration set are required to use

Destination

Represents the destination of the message, consisting of To:, CC:, and BCC:

EventDestination

Contains information about the event destination that the specified email sending

ExtensionField

Additional X-headers to include in the Delivery Status Notification (DSN) when an

IdentityDkimAttributes

Represents the DKIM attributes of a verified email address or a domain.

IdentityMailFromDomainAttributes

Represents the custom MAIL FROM domain attributes of a verified identity (email

IdentityNotificationAttributes

Represents the notification attributes of an identity, including whether an identity

IdentityVerificationAttributes

Represents the verification attributes of a single identity.

KinesisFirehoseDestination

Contains the delivery stream ARN and the IAM role ARN associated with an Amazon Kinesis Firehose event

LambdaAction

When included in a receipt rule, this action calls an AWS Lambda function and,

Message

Represents the message to be sent, composed of a subject and a body.

MessageDsn

Message-related information to include in the Delivery Status Notification (DSN) when

MessageTag

Contains the name and value of a tag that you can provide to SendEmail or

RawMessage

Represents the raw data of the message.

ReceiptAction

An action that Amazon SES can take when it receives an email on behalf of one or more email

ReceiptFilter

A receipt IP address filter enables you to specify whether to accept or reject mail

ReceiptIpFilter

A receipt IP address filter enables you to specify whether to accept or reject mail

ReceiptRule

Receipt rules enable you to specify which actions Amazon SES should take when it receives

ReceiptRuleSetMetadata

Information about a receipt rule set.

RecipientDsnFields

Recipient-related information to include in the Delivery Status Notification (DSN)

ReputationOptions

Contains information about the reputation settings for a configuration set.

S3Action

When included in a receipt rule, this action saves the received message to an Amazon Simple Storage Service

SendDataPoint

Represents sending statistics data. Each SendDataPoint contains

SnsAction

When included in a receipt rule, this action publishes a notification to Amazon Simple Notification Service

SnsDestination

Contains the topic ARN associated with an Amazon Simple Notification Service (Amazon SNS) event destination.

StopAction

When included in a receipt rule, this action terminates the evaluation of the receipt

Template

The content of the email, composed of a subject line, an HTML part, and a text-only

TemplateMetadata

Contains information about an email template.

TrackingOptions

A domain that is used to redirect email recipients to an Amazon SES-operated domain. This

WorkmailAction

When included in a receipt rule, this action calls Amazon WorkMail and, optionally,

Enums

BehaviorOnMxFailure

BounceType

BulkEmailStatus

ConfigurationSetAttribute

CustomMailFromStatus

DimensionValueSource

DsnAction

EventType

IdentityType

InvocationType

NotificationType

ReceiptFilterPolicy

SnsActionEncoding

StopScope

TlsPolicy

VerificationStatus