Module aws_sdk_ses::operation

Structs

CloneReceiptRuleSet

Creates a receipt rule set by cloning an existing one. All receipt rules and

CreateConfigurationSet

Creates a configuration set.

CreateConfigurationSetEventDestination

Creates a configuration set event destination.

CreateConfigurationSetTrackingOptions

Creates an association between a configuration set and a custom domain for open and

CreateCustomVerificationEmailTemplate

Creates a new custom verification email template.

CreateReceiptFilter

Creates a new IP address filter.

CreateReceiptRule

Creates a receipt rule.

CreateReceiptRuleSet

Creates an empty receipt rule set.

CreateTemplate

Creates an email template. Email templates enable you to send personalized email to

DeleteConfigurationSet

Deletes a configuration set. Configuration sets enable you to publish email sending

DeleteConfigurationSetEventDestination

Deletes a configuration set event destination. Configuration set event destinations

DeleteConfigurationSetTrackingOptions

Deletes an association between a configuration set and a custom domain for open and

DeleteCustomVerificationEmailTemplate

Deletes an existing custom verification email template.

DeleteIdentity

Deletes the specified identity (an email address or a domain) from the list of

DeleteIdentityPolicy

Deletes the specified sending authorization policy for the given identity (an email

DeleteReceiptFilter

Deletes the specified IP address filter.

DeleteReceiptRule

Deletes the specified receipt rule.

DeleteReceiptRuleSet

Deletes the specified receipt rule set and all of the receipt rules it

DeleteTemplate

Deletes an email template.

DeleteVerifiedEmailAddress

Deprecated. Use the DeleteIdentity operation to delete email addresses

DescribeActiveReceiptRuleSet

Returns the metadata and receipt rules for the receipt rule set that is currently

DescribeConfigurationSet

Returns the details of the specified configuration set. For information about using

DescribeReceiptRule

Returns the details of the specified receipt rule.

DescribeReceiptRuleSet

Returns the details of the specified receipt rule set.

GetAccountSendingEnabled

Returns the email sending status of the Amazon SES account for the current region.

GetCustomVerificationEmailTemplate

Returns the custom email verification template for the template name you

GetIdentityDkimAttributes

Returns the current status of Easy DKIM signing for an entity. For domain name

GetIdentityMailFromDomainAttributes

Returns the custom MAIL FROM attributes for a list of identities (email addresses :

GetIdentityNotificationAttributes

Given a list of verified identities (email addresses and/or domains), returns a

GetIdentityPolicies

Returns the requested sending authorization policies for the given identity (an email

GetIdentityVerificationAttributes

Given a list of identities (email addresses and/or domains), returns the verification

GetSendQuota

Provides the sending limits for the Amazon SES account.

GetSendStatistics

Provides sending statistics for the current AWS Region. The result is a list of data

GetTemplate

Displays the template object (which includes the Subject line, HTML part and text

ListConfigurationSets

Provides a list of the configuration sets associated with your Amazon SES account in the

ListCustomVerificationEmailTemplates

Lists the existing custom verification email templates for your account in the current

ListIdentities

Returns a list containing all of the identities (email addresses and domains) for your

ListIdentityPolicies

Returns a list of sending authorization policies that are attached to the given

ListReceiptFilters

Lists the IP address filters associated with your AWS account in the current AWS

ListReceiptRuleSets

Lists the receipt rule sets that exist under your AWS account in the current AWS

ListTemplates

Lists the email templates present in your Amazon SES account in the current AWS

ListVerifiedEmailAddresses

Deprecated. Use the ListIdentities operation to list the email addresses

PutConfigurationSetDeliveryOptions

Adds or updates the delivery options for a configuration set.

PutIdentityPolicy

Adds or updates a sending authorization policy for the specified identity (an email

ReorderReceiptRuleSet

Reorders the receipt rules within a receipt rule set.

SendBounce

Generates and sends a bounce message to the sender of an email you received through

SendBulkTemplatedEmail

Composes an email message to multiple destinations. The message body is created using

SendCustomVerificationEmail

Adds an email address to the list of identities for your Amazon SES account in the current

SendEmail

Composes an email message and immediately queues it for sending. In order to send

SendRawEmail

Composes an email message and immediately queues it for sending.

SendTemplatedEmail

Composes an email message using an email template and immediately queues it for

SetActiveReceiptRuleSet

Sets the specified receipt rule set as the active receipt rule set.

SetIdentityDkimEnabled

Enables or disables Easy DKIM signing of email sent from an identity. If Easy DKIM

SetIdentityFeedbackForwardingEnabled

Given an identity (an email address or a domain), enables or disables whether Amazon SES

SetIdentityHeadersInNotificationsEnabled

Given an identity (an email address or a domain), sets whether Amazon SES includes the

SetIdentityMailFromDomain

Enables or disables the custom MAIL FROM domain setup for a verified identity (an

SetIdentityNotificationTopic

Sets an Amazon Simple Notification Service (Amazon SNS) topic to use when delivering notifications. When you use

SetReceiptRulePosition

Sets the position of the specified receipt rule in the receipt rule set.

TestRenderTemplate

Creates a preview of the MIME content of an email when provided with a template and a

UpdateAccountSendingEnabled

Enables or disables email sending across your entire Amazon SES account in the current

UpdateConfigurationSetEventDestination

Updates the event destination of a configuration set. Event destinations are

UpdateConfigurationSetReputationMetricsEnabled

Enables or disables the publishing of reputation metrics for emails sent using a

UpdateConfigurationSetSendingEnabled

Enables or disables email sending for messages sent using a specific configuration set

UpdateConfigurationSetTrackingOptions

Modifies an association between a configuration set and a custom domain for open and

UpdateCustomVerificationEmailTemplate

Updates an existing custom verification email template.

UpdateReceiptRule

Updates a receipt rule.

UpdateTemplate

Updates an email template. Email templates enable you to send personalized email to

VerifyDomainDkim

Returns a set of DKIM tokens for a domain identity.

VerifyDomainIdentity

Adds a domain to the list of identities for your Amazon SES account in the current AWS

VerifyEmailAddress

Deprecated. Use the VerifyEmailIdentity operation to verify a new email

VerifyEmailIdentity

Adds an email address to the list of identities for your Amazon SES account in the current