Module aws_sdk_ses::output

Modules

clone_receipt_rule_set_output

See CloneReceiptRuleSetOutput

create_configuration_set_event_destination_output

See CreateConfigurationSetEventDestinationOutput

create_configuration_set_output

See CreateConfigurationSetOutput

create_configuration_set_tracking_options_output

See CreateConfigurationSetTrackingOptionsOutput

create_custom_verification_email_template_output

See CreateCustomVerificationEmailTemplateOutput

create_receipt_filter_output

See CreateReceiptFilterOutput

create_receipt_rule_output

See CreateReceiptRuleOutput

create_receipt_rule_set_output

See CreateReceiptRuleSetOutput

create_template_output

See CreateTemplateOutput

delete_configuration_set_event_destination_output

See DeleteConfigurationSetEventDestinationOutput

delete_configuration_set_output

See DeleteConfigurationSetOutput

delete_configuration_set_tracking_options_output

See DeleteConfigurationSetTrackingOptionsOutput

delete_custom_verification_email_template_output

See DeleteCustomVerificationEmailTemplateOutput

delete_identity_output

See DeleteIdentityOutput

delete_identity_policy_output

See DeleteIdentityPolicyOutput

delete_receipt_filter_output

See DeleteReceiptFilterOutput

delete_receipt_rule_output

See DeleteReceiptRuleOutput

delete_receipt_rule_set_output

See DeleteReceiptRuleSetOutput

delete_template_output

See DeleteTemplateOutput

delete_verified_email_address_output

See DeleteVerifiedEmailAddressOutput

describe_active_receipt_rule_set_output

See DescribeActiveReceiptRuleSetOutput

describe_configuration_set_output

See DescribeConfigurationSetOutput

describe_receipt_rule_output

See DescribeReceiptRuleOutput

describe_receipt_rule_set_output

See DescribeReceiptRuleSetOutput

get_account_sending_enabled_output

See GetAccountSendingEnabledOutput

get_custom_verification_email_template_output

See GetCustomVerificationEmailTemplateOutput

get_identity_dkim_attributes_output

See GetIdentityDkimAttributesOutput

get_identity_mail_from_domain_attributes_output

See GetIdentityMailFromDomainAttributesOutput

get_identity_notification_attributes_output

See GetIdentityNotificationAttributesOutput

get_identity_policies_output

See GetIdentityPoliciesOutput

get_identity_verification_attributes_output

See GetIdentityVerificationAttributesOutput

get_send_quota_output

See GetSendQuotaOutput

get_send_statistics_output

See GetSendStatisticsOutput

get_template_output

See GetTemplateOutput

list_configuration_sets_output

See ListConfigurationSetsOutput

list_custom_verification_email_templates_output

See ListCustomVerificationEmailTemplatesOutput

list_identities_output

See ListIdentitiesOutput

list_identity_policies_output

See ListIdentityPoliciesOutput

list_receipt_filters_output

See ListReceiptFiltersOutput

list_receipt_rule_sets_output

See ListReceiptRuleSetsOutput

list_templates_output

See ListTemplatesOutput

list_verified_email_addresses_output

See ListVerifiedEmailAddressesOutput

put_configuration_set_delivery_options_output

See PutConfigurationSetDeliveryOptionsOutput

put_identity_policy_output

See PutIdentityPolicyOutput

reorder_receipt_rule_set_output

See ReorderReceiptRuleSetOutput

send_bounce_output

See SendBounceOutput

send_bulk_templated_email_output

See SendBulkTemplatedEmailOutput

send_custom_verification_email_output

See SendCustomVerificationEmailOutput

send_email_output

See SendEmailOutput

send_raw_email_output

See SendRawEmailOutput

send_templated_email_output

See SendTemplatedEmailOutput

set_active_receipt_rule_set_output

See SetActiveReceiptRuleSetOutput

set_identity_dkim_enabled_output

See SetIdentityDkimEnabledOutput

set_identity_feedback_forwarding_enabled_output

See SetIdentityFeedbackForwardingEnabledOutput

set_identity_headers_in_notifications_enabled_output

See SetIdentityHeadersInNotificationsEnabledOutput

set_identity_mail_from_domain_output

See SetIdentityMailFromDomainOutput

set_identity_notification_topic_output

See SetIdentityNotificationTopicOutput

set_receipt_rule_position_output

See SetReceiptRulePositionOutput

test_render_template_output

See TestRenderTemplateOutput

update_account_sending_enabled_output

See UpdateAccountSendingEnabledOutput

update_configuration_set_event_destination_output

See UpdateConfigurationSetEventDestinationOutput

update_configuration_set_reputation_metrics_enabled_output

See UpdateConfigurationSetReputationMetricsEnabledOutput

update_configuration_set_sending_enabled_output

See UpdateConfigurationSetSendingEnabledOutput

update_configuration_set_tracking_options_output

See UpdateConfigurationSetTrackingOptionsOutput

update_custom_verification_email_template_output

See UpdateCustomVerificationEmailTemplateOutput

update_receipt_rule_output

See UpdateReceiptRuleOutput

update_template_output

See UpdateTemplateOutput

verify_domain_dkim_output

See VerifyDomainDkimOutput

verify_domain_identity_output

See VerifyDomainIdentityOutput

verify_email_address_output

See VerifyEmailAddressOutput

verify_email_identity_output

See VerifyEmailIdentityOutput

Structs

CloneReceiptRuleSetOutput

An empty element returned on a successful request.

CreateConfigurationSetEventDestinationOutput

An empty element returned on a successful request.

CreateConfigurationSetOutput

An empty element returned on a successful request.

CreateConfigurationSetTrackingOptionsOutput

An empty element returned on a successful request.

CreateCustomVerificationEmailTemplateOutput

CreateReceiptFilterOutput

An empty element returned on a successful request.

CreateReceiptRuleOutput

An empty element returned on a successful request.

CreateReceiptRuleSetOutput

An empty element returned on a successful request.

CreateTemplateOutput

DeleteConfigurationSetEventDestinationOutput

An empty element returned on a successful request.

DeleteConfigurationSetOutput

An empty element returned on a successful request.

DeleteConfigurationSetTrackingOptionsOutput

An empty element returned on a successful request.

DeleteCustomVerificationEmailTemplateOutput

DeleteIdentityOutput

An empty element returned on a successful request.

DeleteIdentityPolicyOutput

An empty element returned on a successful request.

DeleteReceiptFilterOutput

An empty element returned on a successful request.

DeleteReceiptRuleOutput

An empty element returned on a successful request.

DeleteReceiptRuleSetOutput

An empty element returned on a successful request.

DeleteTemplateOutput

DeleteVerifiedEmailAddressOutput

DescribeActiveReceiptRuleSetOutput

Represents the metadata and receipt rules for the receipt rule set that is currently

DescribeConfigurationSetOutput

Represents the details of a configuration set. Configuration sets enable you to

DescribeReceiptRuleOutput

Represents the details of a receipt rule.

DescribeReceiptRuleSetOutput

Represents the details of the specified receipt rule set.

GetAccountSendingEnabledOutput

Represents a request to return the email sending status for your Amazon SES account in the

GetCustomVerificationEmailTemplateOutput

The content of the custom verification email template.

GetIdentityDkimAttributesOutput

Represents the status of Amazon SES Easy DKIM signing for an identity. For domain

GetIdentityMailFromDomainAttributesOutput

Represents the custom MAIL FROM attributes for a list of identities.

GetIdentityNotificationAttributesOutput

Represents the notification attributes for a list of identities.

GetIdentityPoliciesOutput

Represents the requested sending authorization policies.

GetIdentityVerificationAttributesOutput

The Amazon SES verification status of a list of identities. For domain identities, this

GetSendQuotaOutput

Represents your Amazon SES daily sending quota, maximum send rate, and the number of emails

GetSendStatisticsOutput

Represents a list of data points. This list contains aggregated data from the previous

GetTemplateOutput

ListConfigurationSetsOutput

A list of configuration sets associated with your AWS account. Configuration sets

ListCustomVerificationEmailTemplatesOutput

A paginated list of custom verification email templates.

ListIdentitiesOutput

A list of all identities that you have attempted to verify under your AWS account,

ListIdentityPoliciesOutput

A list of names of sending authorization policies that apply to an identity.

ListReceiptFiltersOutput

A list of IP address filters that exist under your AWS account.

ListReceiptRuleSetsOutput

A list of receipt rule sets that exist under your AWS account.

ListTemplatesOutput

ListVerifiedEmailAddressesOutput

A list of email addresses that you have verified with Amazon SES under your AWS

PutConfigurationSetDeliveryOptionsOutput

An HTTP 200 response if the request succeeds, or an error message if the request

PutIdentityPolicyOutput

An empty element returned on a successful request.

ReorderReceiptRuleSetOutput

An empty element returned on a successful request.

SendBounceOutput

Represents a unique message ID.

SendBulkTemplatedEmailOutput

SendCustomVerificationEmailOutput

The response received when attempting to send the custom verification email.

SendEmailOutput

Represents a unique message ID.

SendRawEmailOutput

Represents a unique message ID.

SendTemplatedEmailOutput

SetActiveReceiptRuleSetOutput

An empty element returned on a successful request.

SetIdentityDkimEnabledOutput

An empty element returned on a successful request.

SetIdentityFeedbackForwardingEnabledOutput

An empty element returned on a successful request.

SetIdentityHeadersInNotificationsEnabledOutput

An empty element returned on a successful request.

SetIdentityMailFromDomainOutput

An empty element returned on a successful request.

SetIdentityNotificationTopicOutput

An empty element returned on a successful request.

SetReceiptRulePositionOutput

An empty element returned on a successful request.

TestRenderTemplateOutput

UpdateAccountSendingEnabledOutput

UpdateConfigurationSetEventDestinationOutput

An empty element returned on a successful request.

UpdateConfigurationSetReputationMetricsEnabledOutput

UpdateConfigurationSetSendingEnabledOutput

UpdateConfigurationSetTrackingOptionsOutput

An empty element returned on a successful request.

UpdateCustomVerificationEmailTemplateOutput

UpdateReceiptRuleOutput

An empty element returned on a successful request.

UpdateTemplateOutput

VerifyDomainDkimOutput

Returns CNAME records that you must publish to the DNS server of your domain to set up

VerifyDomainIdentityOutput

Returns a TXT record that you must publish to the DNS server of your domain to

VerifyEmailAddressOutput

VerifyEmailIdentityOutput

An empty element returned on a successful request.