Struct aws_sdk_ses::client::Client

pub struct Client<C = DynConnector, M = AwsMiddleware, R = Standard> { /* fields omitted */ }

An ergonomic service client for SimpleEmailService.

This client allows ergonomic access to a SimpleEmailService-shaped service. Each method corresponds to an endpoint defined in the service’s Smithy model, and the request and response shapes are auto-generated from that same model.

Using a Client

Once you have a client set up, you can access the service’s endpoints by calling the appropriate method on Client. Each such method returns a request builder for that endpoint, with methods for setting the various fields of the request. Once your request is complete, use the send method to send the request. send returns a future, which you then have to .await to get the service’s response.

Implementations

impl<C, M, R> Client<C, M, R>

pub fn with_config(client: Client<C, M, R>, conf: Config) -> Self

Creates a client with the given service configuration.

pub fn conf(&self) -> &Config

Returns the client’s configuration.

impl<C, M, R> Client<C, M, R> where
C: SmithyConnector,
M: SmithyMiddleware<C>,
R: NewRequestPolicy,

pub fn clone_receipt_rule_set(&self) -> CloneReceiptRuleSet<C, M, R>

Constructs a fluent builder for the CloneReceiptRuleSet operation.

See CloneReceiptRuleSet for more information about the operation and its arguments.

pub fn create_configuration_set(&self) -> CreateConfigurationSet<C, M, R>

Constructs a fluent builder for the CreateConfigurationSet operation.

See CreateConfigurationSet for more information about the operation and its arguments.

pub fn create_configuration_set_event_destination(
&self
) -> CreateConfigurationSetEventDestination<C, M, R>

Constructs a fluent builder for the CreateConfigurationSetEventDestination operation.

See CreateConfigurationSetEventDestination for more information about the operation and its arguments.

pub fn create_configuration_set_tracking_options(
&self
) -> CreateConfigurationSetTrackingOptions<C, M, R>

Constructs a fluent builder for the CreateConfigurationSetTrackingOptions operation.

See CreateConfigurationSetTrackingOptions for more information about the operation and its arguments.

pub fn create_custom_verification_email_template(
&self
) -> CreateCustomVerificationEmailTemplate<C, M, R>

Constructs a fluent builder for the CreateCustomVerificationEmailTemplate operation.

See CreateCustomVerificationEmailTemplate for more information about the operation and its arguments.

pub fn create_receipt_filter(&self) -> CreateReceiptFilter<C, M, R>

Constructs a fluent builder for the CreateReceiptFilter operation.

See CreateReceiptFilter for more information about the operation and its arguments.

pub fn create_receipt_rule(&self) -> CreateReceiptRule<C, M, R>

Constructs a fluent builder for the CreateReceiptRule operation.

See CreateReceiptRule for more information about the operation and its arguments.

pub fn create_receipt_rule_set(&self) -> CreateReceiptRuleSet<C, M, R>

Constructs a fluent builder for the CreateReceiptRuleSet operation.

See CreateReceiptRuleSet for more information about the operation and its arguments.

pub fn create_template(&self) -> CreateTemplate<C, M, R>

Constructs a fluent builder for the CreateTemplate operation.

See CreateTemplate for more information about the operation and its arguments.

pub fn delete_configuration_set(&self) -> DeleteConfigurationSet<C, M, R>

Constructs a fluent builder for the DeleteConfigurationSet operation.

See DeleteConfigurationSet for more information about the operation and its arguments.

pub fn delete_configuration_set_event_destination(
&self
) -> DeleteConfigurationSetEventDestination<C, M, R>

Constructs a fluent builder for the DeleteConfigurationSetEventDestination operation.

See DeleteConfigurationSetEventDestination for more information about the operation and its arguments.

pub fn delete_configuration_set_tracking_options(
&self
) -> DeleteConfigurationSetTrackingOptions<C, M, R>

Constructs a fluent builder for the DeleteConfigurationSetTrackingOptions operation.

See DeleteConfigurationSetTrackingOptions for more information about the operation and its arguments.

pub fn delete_custom_verification_email_template(
&self
) -> DeleteCustomVerificationEmailTemplate<C, M, R>

Constructs a fluent builder for the DeleteCustomVerificationEmailTemplate operation.

See DeleteCustomVerificationEmailTemplate for more information about the operation and its arguments.

pub fn delete_identity(&self) -> DeleteIdentity<C, M, R>

Constructs a fluent builder for the DeleteIdentity operation.

See DeleteIdentity for more information about the operation and its arguments.

pub fn delete_identity_policy(&self) -> DeleteIdentityPolicy<C, M, R>

Constructs a fluent builder for the DeleteIdentityPolicy operation.

See DeleteIdentityPolicy for more information about the operation and its arguments.

pub fn delete_receipt_filter(&self) -> DeleteReceiptFilter<C, M, R>

Constructs a fluent builder for the DeleteReceiptFilter operation.

See DeleteReceiptFilter for more information about the operation and its arguments.

pub fn delete_receipt_rule(&self) -> DeleteReceiptRule<C, M, R>

Constructs a fluent builder for the DeleteReceiptRule operation.

See DeleteReceiptRule for more information about the operation and its arguments.

pub fn delete_receipt_rule_set(&self) -> DeleteReceiptRuleSet<C, M, R>

Constructs a fluent builder for the DeleteReceiptRuleSet operation.

See DeleteReceiptRuleSet for more information about the operation and its arguments.

pub fn delete_template(&self) -> DeleteTemplate<C, M, R>

Constructs a fluent builder for the DeleteTemplate operation.

See DeleteTemplate for more information about the operation and its arguments.

pub fn delete_verified_email_address(
&self
) -> DeleteVerifiedEmailAddress<C, M, R>

Constructs a fluent builder for the DeleteVerifiedEmailAddress operation.

See DeleteVerifiedEmailAddress for more information about the operation and its arguments.

pub fn describe_active_receipt_rule_set(
&self
) -> DescribeActiveReceiptRuleSet<C, M, R>

Constructs a fluent builder for the DescribeActiveReceiptRuleSet operation.

See DescribeActiveReceiptRuleSet for more information about the operation and its arguments.

pub fn describe_configuration_set(&self) -> DescribeConfigurationSet<C, M, R>

Constructs a fluent builder for the DescribeConfigurationSet operation.

See DescribeConfigurationSet for more information about the operation and its arguments.

pub fn describe_receipt_rule(&self) -> DescribeReceiptRule<C, M, R>

Constructs a fluent builder for the DescribeReceiptRule operation.

See DescribeReceiptRule for more information about the operation and its arguments.

pub fn describe_receipt_rule_set(&self) -> DescribeReceiptRuleSet<C, M, R>

Constructs a fluent builder for the DescribeReceiptRuleSet operation.

See DescribeReceiptRuleSet for more information about the operation and its arguments.

pub fn get_account_sending_enabled(&self) -> GetAccountSendingEnabled<C, M, R>

Constructs a fluent builder for the GetAccountSendingEnabled operation.

See GetAccountSendingEnabled for more information about the operation and its arguments.

pub fn get_custom_verification_email_template(
&self
) -> GetCustomVerificationEmailTemplate<C, M, R>

Constructs a fluent builder for the GetCustomVerificationEmailTemplate operation.

See GetCustomVerificationEmailTemplate for more information about the operation and its arguments.

pub fn get_identity_dkim_attributes(&self) -> GetIdentityDkimAttributes<C, M, R>

Constructs a fluent builder for the GetIdentityDkimAttributes operation.

See GetIdentityDkimAttributes for more information about the operation and its arguments.

pub fn get_identity_mail_from_domain_attributes(
&self
) -> GetIdentityMailFromDomainAttributes<C, M, R>

Constructs a fluent builder for the GetIdentityMailFromDomainAttributes operation.

See GetIdentityMailFromDomainAttributes for more information about the operation and its arguments.

pub fn get_identity_notification_attributes(
&self
) -> GetIdentityNotificationAttributes<C, M, R>

Constructs a fluent builder for the GetIdentityNotificationAttributes operation.

See GetIdentityNotificationAttributes for more information about the operation and its arguments.

pub fn get_identity_policies(&self) -> GetIdentityPolicies<C, M, R>

Constructs a fluent builder for the GetIdentityPolicies operation.

See GetIdentityPolicies for more information about the operation and its arguments.

pub fn get_identity_verification_attributes(
&self
) -> GetIdentityVerificationAttributes<C, M, R>

Constructs a fluent builder for the GetIdentityVerificationAttributes operation.

See GetIdentityVerificationAttributes for more information about the operation and its arguments.

pub fn get_send_quota(&self) -> GetSendQuota<C, M, R>

Constructs a fluent builder for the GetSendQuota operation.

See GetSendQuota for more information about the operation and its arguments.

pub fn get_send_statistics(&self) -> GetSendStatistics<C, M, R>

Constructs a fluent builder for the GetSendStatistics operation.

See GetSendStatistics for more information about the operation and its arguments.

pub fn get_template(&self) -> GetTemplate<C, M, R>

Constructs a fluent builder for the GetTemplate operation.

See GetTemplate for more information about the operation and its arguments.

pub fn list_configuration_sets(&self) -> ListConfigurationSets<C, M, R>

Constructs a fluent builder for the ListConfigurationSets operation.

See ListConfigurationSets for more information about the operation and its arguments.

pub fn list_custom_verification_email_templates(
&self
) -> ListCustomVerificationEmailTemplates<C, M, R>

Constructs a fluent builder for the ListCustomVerificationEmailTemplates operation.

See ListCustomVerificationEmailTemplates for more information about the operation and its arguments.

pub fn list_identities(&self) -> ListIdentities<C, M, R>

Constructs a fluent builder for the ListIdentities operation.

See ListIdentities for more information about the operation and its arguments.

pub fn list_identity_policies(&self) -> ListIdentityPolicies<C, M, R>

Constructs a fluent builder for the ListIdentityPolicies operation.

See ListIdentityPolicies for more information about the operation and its arguments.

pub fn list_receipt_filters(&self) -> ListReceiptFilters<C, M, R>

Constructs a fluent builder for the ListReceiptFilters operation.

See ListReceiptFilters for more information about the operation and its arguments.

pub fn list_receipt_rule_sets(&self) -> ListReceiptRuleSets<C, M, R>

Constructs a fluent builder for the ListReceiptRuleSets operation.

See ListReceiptRuleSets for more information about the operation and its arguments.

pub fn list_templates(&self) -> ListTemplates<C, M, R>

Constructs a fluent builder for the ListTemplates operation.

See ListTemplates for more information about the operation and its arguments.

pub fn list_verified_email_addresses(
&self
) -> ListVerifiedEmailAddresses<C, M, R>

Constructs a fluent builder for the ListVerifiedEmailAddresses operation.

See ListVerifiedEmailAddresses for more information about the operation and its arguments.

pub fn put_configuration_set_delivery_options(
&self
) -> PutConfigurationSetDeliveryOptions<C, M, R>

Constructs a fluent builder for the PutConfigurationSetDeliveryOptions operation.

See PutConfigurationSetDeliveryOptions for more information about the operation and its arguments.

pub fn put_identity_policy(&self) -> PutIdentityPolicy<C, M, R>

Constructs a fluent builder for the PutIdentityPolicy operation.

See PutIdentityPolicy for more information about the operation and its arguments.

pub fn reorder_receipt_rule_set(&self) -> ReorderReceiptRuleSet<C, M, R>

Constructs a fluent builder for the ReorderReceiptRuleSet operation.

See ReorderReceiptRuleSet for more information about the operation and its arguments.

pub fn send_bounce(&self) -> SendBounce<C, M, R>

Constructs a fluent builder for the SendBounce operation.

See SendBounce for more information about the operation and its arguments.

pub fn send_bulk_templated_email(&self) -> SendBulkTemplatedEmail<C, M, R>

Constructs a fluent builder for the SendBulkTemplatedEmail operation.

See SendBulkTemplatedEmail for more information about the operation and its arguments.

pub fn send_custom_verification_email(
&self
) -> SendCustomVerificationEmail<C, M, R>

Constructs a fluent builder for the SendCustomVerificationEmail operation.

See SendCustomVerificationEmail for more information about the operation and its arguments.

pub fn send_email(&self) -> SendEmail<C, M, R>

Constructs a fluent builder for the SendEmail operation.

See SendEmail for more information about the operation and its arguments.

pub fn send_raw_email(&self) -> SendRawEmail<C, M, R>

Constructs a fluent builder for the SendRawEmail operation.

See SendRawEmail for more information about the operation and its arguments.

pub fn send_templated_email(&self) -> SendTemplatedEmail<C, M, R>

Constructs a fluent builder for the SendTemplatedEmail operation.

See SendTemplatedEmail for more information about the operation and its arguments.

pub fn set_active_receipt_rule_set(&self) -> SetActiveReceiptRuleSet<C, M, R>

Constructs a fluent builder for the SetActiveReceiptRuleSet operation.

See SetActiveReceiptRuleSet for more information about the operation and its arguments.

pub fn set_identity_dkim_enabled(&self) -> SetIdentityDkimEnabled<C, M, R>

Constructs a fluent builder for the SetIdentityDkimEnabled operation.

See SetIdentityDkimEnabled for more information about the operation and its arguments.

pub fn set_identity_feedback_forwarding_enabled(
&self
) -> SetIdentityFeedbackForwardingEnabled<C, M, R>

Constructs a fluent builder for the SetIdentityFeedbackForwardingEnabled operation.

See SetIdentityFeedbackForwardingEnabled for more information about the operation and its arguments.

pub fn set_identity_headers_in_notifications_enabled(
&self
) -> SetIdentityHeadersInNotificationsEnabled<C, M, R>

Constructs a fluent builder for the SetIdentityHeadersInNotificationsEnabled operation.

See SetIdentityHeadersInNotificationsEnabled for more information about the operation and its arguments.

pub fn set_identity_mail_from_domain(
&self
) -> SetIdentityMailFromDomain<C, M, R>

Constructs a fluent builder for the SetIdentityMailFromDomain operation.

See SetIdentityMailFromDomain for more information about the operation and its arguments.

pub fn set_identity_notification_topic(
&self
) -> SetIdentityNotificationTopic<C, M, R>

Constructs a fluent builder for the SetIdentityNotificationTopic operation.

See SetIdentityNotificationTopic for more information about the operation and its arguments.

pub fn set_receipt_rule_position(&self) -> SetReceiptRulePosition<C, M, R>

Constructs a fluent builder for the SetReceiptRulePosition operation.

See SetReceiptRulePosition for more information about the operation and its arguments.

pub fn test_render_template(&self) -> TestRenderTemplate<C, M, R>

Constructs a fluent builder for the TestRenderTemplate operation.

See TestRenderTemplate for more information about the operation and its arguments.

pub fn update_account_sending_enabled(
&self
) -> UpdateAccountSendingEnabled<C, M, R>

Constructs a fluent builder for the UpdateAccountSendingEnabled operation.

See UpdateAccountSendingEnabled for more information about the operation and its arguments.

pub fn update_configuration_set_event_destination(
&self
) -> UpdateConfigurationSetEventDestination<C, M, R>

Constructs a fluent builder for the UpdateConfigurationSetEventDestination operation.

See UpdateConfigurationSetEventDestination for more information about the operation and its arguments.

pub fn update_configuration_set_reputation_metrics_enabled(
&self
) -> UpdateConfigurationSetReputationMetricsEnabled<C, M, R>

Constructs a fluent builder for the UpdateConfigurationSetReputationMetricsEnabled operation.

See UpdateConfigurationSetReputationMetricsEnabled for more information about the operation and its arguments.

pub fn update_configuration_set_sending_enabled(
&self
) -> UpdateConfigurationSetSendingEnabled<C, M, R>

Constructs a fluent builder for the UpdateConfigurationSetSendingEnabled operation.

See UpdateConfigurationSetSendingEnabled for more information about the operation and its arguments.

pub fn update_configuration_set_tracking_options(
&self
) -> UpdateConfigurationSetTrackingOptions<C, M, R>

Constructs a fluent builder for the UpdateConfigurationSetTrackingOptions operation.

See UpdateConfigurationSetTrackingOptions for more information about the operation and its arguments.

pub fn update_custom_verification_email_template(
&self
) -> UpdateCustomVerificationEmailTemplate<C, M, R>

Constructs a fluent builder for the UpdateCustomVerificationEmailTemplate operation.

See UpdateCustomVerificationEmailTemplate for more information about the operation and its arguments.

pub fn update_receipt_rule(&self) -> UpdateReceiptRule<C, M, R>

Constructs a fluent builder for the UpdateReceiptRule operation.

See UpdateReceiptRule for more information about the operation and its arguments.

pub fn update_template(&self) -> UpdateTemplate<C, M, R>

Constructs a fluent builder for the UpdateTemplate operation.

See UpdateTemplate for more information about the operation and its arguments.

pub fn verify_domain_dkim(&self) -> VerifyDomainDkim<C, M, R>

Constructs a fluent builder for the VerifyDomainDkim operation.

See VerifyDomainDkim for more information about the operation and its arguments.

pub fn verify_domain_identity(&self) -> VerifyDomainIdentity<C, M, R>

Constructs a fluent builder for the VerifyDomainIdentity operation.

See VerifyDomainIdentity for more information about the operation and its arguments.

pub fn verify_email_address(&self) -> VerifyEmailAddress<C, M, R>

Constructs a fluent builder for the VerifyEmailAddress operation.

See VerifyEmailAddress for more information about the operation and its arguments.

pub fn verify_email_identity(&self) -> VerifyEmailIdentity<C, M, R>

Constructs a fluent builder for the VerifyEmailIdentity operation.

See VerifyEmailIdentity for more information about the operation and its arguments.

impl<C> Client<C, AwsMiddleware, Standard>

pub fn from_conf_conn(conf: Config, conn: C) -> Self

Creates a client with the given service config and connector override.

impl Client<DynConnector, AwsMiddleware, Standard>

pub fn new(config: &Config) -> Self

Creates a new client from a shared config.

pub fn from_conf(conf: Config) -> Self

Creates a new client from the service Config.